gnawesome cleanup right_panel to floating_panel var
