+++ config/awesome/gnawesome/layout/floating-panel/floating-panel-opener.lua
@@ -1,0 +1,120 @@
+local awful = require('awful')
+local wibox = require('wibox')
+local gears = require('gears')
+local beautiful = require('beautiful')
+
+local dpi = require('beautiful').xresources.apply_dpi
+local clickable_container = require('widget.clickable-container')
+
+local config_dir = gears.filesystem.get_configuration_dir()
+local widget_icon_dir = config_dir .. 'layout/floating-panel/icons/'
+
+--   ▄▄▄▄▄           ▄      ▄                 
+--   █    █ ▄   ▄  ▄▄█▄▄  ▄▄█▄▄   ▄▄▄   ▄ ▄▄  
+--   █▄▄▄▄▀ █   █    █      █    █▀ ▀█  █▀  █ 
+--   █    █ █   █    █      █    █   █  █   █ 
+--   █▄▄▄▄▀ ▀▄▄▀█    ▀▄▄    ▀▄▄  ▀█▄█▀  █   █ 
+
+
+-- The button in top panel
+
+local return_button = function()
+
+	local widget =
+		wibox.widget {
+		{
+			id = 'icon',
+			image = widget_icon_dir .. 'notification' .. '.svg',
+			widget = wibox.widget.imagebox,
+			resize = true
+		},
+		layout = wibox.layout.align.horizontal
+	}
+
+	local clock_widget = wibox.widget.textclock(
+		'<span font="SF Pro Text Bold 11">%l:%M %p</span>',
+		1
+	)
+
+	local widget_button = wibox.widget {
+		{
+			clock_widget,
+			margins = dpi(7),
+			widget = wibox.container.margin
+		},
+		widget = clickable_container
+	}
+
+	widget_button:buttons(
+		gears.table.join(
+			awful.button(
+				{},
+				1,
+				nil,
+				function()
+					awful.screen.focused().floating_panel:toggle()
+				end
+			)
+		)
+	)
+
+
+	local clock_tooltip = awful.tooltip
+	{
+		objects = {widget_button},
+		mode = 'outside',
+		delay_show = 1,
+		preferred_positions = {'right', 'left', 'top', 'bottom'},
+		preferred_alignments = {'middle'},
+		margin_leftright = dpi(8),
+		margin_topbottom = dpi(8),
+		timer_function = function()
+			local ordinal = nil
+
+			local day = os.date('%d')
+			local month = os.date('%B')
+
+			local first_digit = string.sub(day, 0, 1) 
+			local last_digit = string.sub(day, -1) 
+
+			if first_digit == '0' then
+			  day = last_digit
+			end
+
+
+			if last_digit == '1' and day ~= '11' then
+			  ordinal = 'st'
+			elseif last_digit == '2' and day ~= '12' then
+			  ordinal = 'nd'
+			elseif last_digit == '3' and day ~= '13' then
+			  ordinal = 'rd'
+			else
+			  ordinal = 'th'
+			end
+
+			local date_str = 'Today is the ' ..
+			'<b>' .. day .. ordinal .. 
+			' of ' .. month .. '</b>.\n' ..
+			'And it\'s fucking ' .. os.date('%A')
+
+			return date_str
+
+		end,
+	}
+	
+	widget_button:connect_signal(
+		'button::press', 
+		function(self, lx, ly, button)
+			-- Hide the tooltip when you press the clock widget
+			if clock_tooltip.visible and button == 1 then
+				clock_tooltip.visible = false
+			end
+		end
+	)
+
+	return widget_button
+
+end
+
+
+return return_button
