+++ config/awesome/gnawesome/module/notifications.lua
@@ -253,9 +253,9 @@ naughty.connect_signal("request::display", function(n)
 	}
 
 	-- Destroy popups if dont_disturb mode is on
-	-- Or if the right_panel is visible
+	-- Or if the floating_panel is visible
 	local focused = awful.screen.focused()
-	if _G.dont_disturb or (focused.right_panel and focused.right_panel.visible) then
+	if _G.dont_disturb or (focused.floating_panel and focused.floating_panel.visible) then
 		naughty.destroy_all_notifications(nil, 1)
 	end
